+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # scorphan
 
+## [0.4.0] - 2024/08/29
+
+### Changed
+
+- Added an option to choose between scrublet and vaeda for doublet detection
+- Moved doublet detection to after removal of low quality cells (as per https://www.sc-best-practices.org/)
+
 ## [0.3.0] - 2024/08/09
 
 ### Changed